image_metadata/requirements.txt
Pillow
piexif
pillow-heif
requests 